[feat] 04/02 CommunityPostHeaderContainer 분리 #422
